// generated/src/app/api-models/claim-hold-deter-values.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ClaimHoldDeterValues {
  seqClhldRule: number;
  determinantColumnNo: number;
  determinantValueNo: number;
  detFromValue: string;
  detThruValue: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}